final updates

+++ gluefactory/models/extractors/sift.py
@@ -221,7 +221,7 @@ class SIFT(BaseModel):
             img = image[k]
             if "image_size" in data.keys():
                 # avoid extracting points in padded areas
-                w, h = data["image_size"][k]
+                w, h = data["image_size"][k].to(int)
                 img = img[:, :h, :w]
             p = self.extract_single_image(img)
             pred.append(p)
